move to Methods dir for backup methods, changing Replset + ShardingHandler to take in DB class instead of make a new one
