+++ MongoBackup/Methods/__init__.py
@@ -1,0 +1,2 @@
+from Mongodump import Mongodump
+from Mongodumper import Mongodumper
